Fix i686 ticket-key wrap, OPENSSL_init_ssl width, and 32-bit test literals
Ported from #157 (Anshul Singhvi) on top of the Win32 stdcall work from #153:

- Session ticket key rotation and lifetime constants multiplied in Int32 on
  i686 and wrapped negative, so every handshake rotated the key and no TLS 1.2
  session could resume ("unexpected resumed state" during precompile).
- OPENSSL_init_ssl takes uint64_t opts; passing Culong shifted the settings
  pointer on 32-bit targets.
- inet_pton is called from Ws2_32 by name on Windows.
- Test literals passed to Int64 parameters are wrapped, and the TLS server task
  closes its socket on any failure so a client read cannot strand the suite.
- CI gains an ubuntu x86 row; the windows-min-tls-debug workflow names a test
  file that exists.

+++ test/tls_test_utils.jl
@@ -42,7 +42,9 @@ end
 end
 
 function _tls_server_config(;
-    handshake_timeout_ns::Int64 = 0,
+    # `Integer`, matching `TL.Config`: typed keywords assert rather than convert,
+    # and literals such as `2_000_000_000` are `Int32` on 32-bit Julia.
+    handshake_timeout_ns::Integer = 0,
     cert_file::String = _TLS_CERT_PATH,
     key_file::String = _TLS_KEY_PATH,
     client_auth::TL.ClientAuthMode.T = TL.ClientAuthMode.NoClientCert,
